moved allans files into mvc folders
