+++ seeker/src/main/java/se452/group9/seeker/repo/ApplicationRepository.java
@@ -1,9 +1,9 @@
-package se452.group9.seeker;
+package se452.group9.seeker.repo;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import se452.group9.seeker.School;
+import se452.group9.seeker.model.School;
 
 public interface ApplicationRepository extends JpaRepository <Application, Long>{
