+++ seeker/src/main/java/se452/group9/seeker/repo/LocationRepository.java
@@ -1,9 +1,9 @@
-package se452.group9.seeker;
+package se452.group9.seeker.model;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import se452.group9.seeker.School;
+import se452.group9.seeker.model.School;
 
 public interface LocationRepository extends JpaRepository <Location, Long>{
